integration-tests/requirements.txt
kubernetes
psycopg2-binary
retry